extract schedule module from scenario
(needed to extract ResourceUtc from utils)
