+++ source/jormungandr/jormungandr/utils.py
@@ -31,11 +31,6 @@ import calendar
 from collections import deque
 from datetime import datetime
 from google.protobuf.descriptor import FieldDescriptor
-from google.protobuf.internal.containers import RepeatedScalarFieldContainer
-from jormungandr import i_manager
-import logging
-import pytz
-from jormungandr.exceptions import RegionNotFound
 from navitiacommon import response_pb2
 
 
@@ -55,60 +50,6 @@ def date_to_timestamp(date):
     convert a datatime objet to a posix timestamp (number of seconds from 1070/1/1)
     """
     return int(calendar.timegm(date.utctimetuple()))
-
-
-class ResourceUtc:
-    def __init__(self):
-        self._tz = None
-
-    def tz(self):
-        if not self._tz:
-            instance = i_manager.instances.get(self.region, None)
-
-            if not instance:
-                raise RegionNotFound(self.region)
-
-            tz_name = instance.timezone  # TODO store directly the tz?
-
-            if not tz_name:
-                logging.Logger(__name__).warn("unknown timezone for region {}"
-                                              .format(self.region))
-                return None
-            self._tz = (pytz.timezone(tz_name),)
-        return self._tz[0]
-
-    def convert_to_utc(self, original_datetime):
-        """
-        convert the original_datetime in the args to UTC
-
-        for that we need to 'guess' the timezone wanted by the user
-
-        For the moment We only use the default instance timezone.
-
-        It won't obviously work for multi timezone instances, we'll have to do
-        something smarter.
-
-        We'll have to consider either the departure or the arrival of the journey
-        (depending on the `clockwise` param)
-        and fetch the tz of this point.
-        we'll have to store the tz for stop area and the coord for admin, poi, ...
-        """
-
-        if self.tz() is None:
-            return original_datetime
-
-        utctime = self.tz().normalize(self.tz().localize(original_datetime)).astimezone(pytz.utc)
-
-        return utctime
-
-    def format(self, value):
-        dt = datetime.utcfromtimestamp(value)
-
-        if self.tz() is not None:
-            dt = pytz.utc.localize(dt)
-            dt = dt.astimezone(self.tz())
-            return dt.strftime("%Y%m%dT%H%M%S")
-        return None  # for the moment I prefer not to display anything instead of something wrong
 
 
 def walk_dict(tree, visitor):
